+++ application/modules/editor/Models/Import/FileParser/XmlParser.php
@@ -74,7 +74,7 @@ class editor_Models_Import_FileParser_XmlParser {
     protected $preserveWhitespace;
     
     protected $nonXmlBlocks = [];
-    
+
     /**
      * walks through the given XML string and fires the registered callbacks for each found node
      * Preserving whitespace in XML is defined by the xml:space attribute on each node.
@@ -82,9 +82,10 @@ class editor_Models_Import_FileParser_XmlParser {
      *  The initial root value (preserve or ignore) is given here as boolean parameter
      * @param string $xml
      * @param bool $preserveWhitespaceRoot
+     * @param array $validTags optional, if given only the listed tags are considered to be tags, all other tag like contents are then considered as text and the entities are ecnoded properly
      * @return string the parsed string with all callbacks applied
      */
-    public function parse($xml, $preserveWhitespaceRoot = false) {
+    public function parse(string $xml, bool $preserveWhitespaceRoot = false, array $validTags = []): string {
         $this->nonXmlBlocks = [];
         $xml = preg_replace_callback('/(<!\[CDATA\[.*?\]\]>)|(<!--.*?-->)/s', function($item){
             $id = count($this->nonXmlBlocks);
@@ -107,7 +108,36 @@ class editor_Models_Import_FileParser_XmlParser {
         //    Element names cannot contain spaces                                       => is implicitly checked on parsing the tag chunk below, since spaces are used as separator for the attributes.
 
         // see also the regex in parseList!
-        $this->parseList(preg_split('#(</?[a-zA-Z_][^>]*>)#i', $xml, flags: PREG_SPLIT_DELIM_CAPTURE), $preserveWhitespaceRoot);
+        if(empty($validTags)) {
+            $regex = '#(</?[a-zA-Z_][^>]*>)#i';
+        }
+        else {
+            $regex = '#(</?('.join('|', $validTags).')[^>]*>)#i';
+        }
+        $chunks = preg_split($regex, $xml, flags: PREG_SPLIT_DELIM_CAPTURE);
+        if(!empty($validTags)) {
+            //with validTags we have a 2nd pair of parentheses, we have to clean the captured content then:
+/*
+ *             Array
+                (
+                    [0] =>
+                    [1] => <foo>
+                    [2] => foo                  → to be removed
+                    [3] =>  & <test &amp;
+                    [4] => <ph type="lb"/>
+                    [5] => ph                   → to be removed
+                    [6] =>
+                    [7] => </foo>
+                    [8] => foo                  → to be removed
+                    [9] =>
+                )
+*/
+            $chunks = array_values(array_filter($chunks, function ($key){
+                //if key modulo 3 is 2, then this is the second regex parenthesis content and can be ignored!
+                return $key % 3 !== 2;
+            },  ARRAY_FILTER_USE_KEY));
+        }
+        $this->parseList($chunks, $preserveWhitespaceRoot);
         return str_replace(array_keys($this->nonXmlBlocks), array_values($this->nonXmlBlocks), $this->__toString());
     }